Make fields a lazy property of Document
